+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Updated to work with Bones `v4.0`
+- Moved deploy config to `config/deploy.php`
 
 ## [1.0.2] - 2023.08.10